Changed the MeetingParams Time Inputs
The Time inputs are back to TextViews and this is because you can make a
TimePicker dialog box that pops up over the activity. This will be nicer
looking than shoving three TimePickers onto the same screen along with
all the other crap. Current the method that displays the TimePicker
requires the associated TextView as an argument. It then changes the
TextView's text. As well, there are now methods for minute parsing and
hour parsing.